requirements.txt
# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
pydantic==2.5.2

# LLM Gateway
groq==0.3.1
python-dotenv==1.0.0

# Vector DB & Embeddings
chromadb==0.4.18
sentence-transformers==2.2.2
torch==2.1.1  # For CPU embeddings

# Scraping & Processing
playwright==1.40.0
beautifulsoup4==4.12.2
pandas==2.1.3
requests==2.31.0

# Testing
pytest==7.4.3
httpx==0.25.1
